fix the field-table encoding and the consumer teardown after review
A non-string key is dropped only at the top level, as the extension drops it, so
a header carrying a list keeps its values; a connection-level failure marks the
connection disconnected; and a consumer's basic.cancel is serialized against the
other commands of its channel instead of racing the channel close, which the
broker answered with 503 and a torn-down connection.

Validation: make check (php-cs-fixer, PHPStan level 6, 635 PHPUnit tests of which
115 are AMQP, go test -race), seven consecutive AMQP suite runs, and the headers
compared byte for byte against ext-amqp in both directions.

+++ ext/internal/features/amqp/payloads/table_test.go
@@ -43,7 +43,9 @@ func TestAMapDecodesIntoTheTable(t *testing.T) {
 		t.Fatalf("table = %#v, want one value", table)
 	}
 
-	if table["x-max-length"] != int8(10) {
+	// Decoded loosely: every integer arrives as an int64, whichever width MessagePack
+	// packed it in.
+	if table["x-max-length"] != int64(10) {
 		t.Fatalf("x-max-length = %#v (%T), want the value that was sent", table["x-max-length"], table["x-max-length"])
 	}
 }
